LayoutTests/svg/animations/svgangle-animation-deg-to-grad-expected.txt

SVG 1.1 dynamic animation tests

Tests SVGAngle animation from deg to grad.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS marker.orientAngle.animVal.value is 0
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientAngle.animVal.value is 90
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientAngle.animVal.value is 180
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientAngle.animVal.value is 0
PASS marker.orientAngle.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
